refactor: put components in view or containers folders
