+++ src/components/containers/DayCard/dayCard.module.scss
@@ -1,4 +1,4 @@
-@import '../../sass/abstract/variables';
+@import '../../../sass/abstract/variables';
 
 .container {
   display: inline-flex;
