+++ src/components/view/Spinner/spinner.module.scss
