+++ src/pages/Login/Login.tsx
@@ -5,7 +5,7 @@ import { logInWithEmailAndPassword, signInWithGoogle } from "../../firebase";
 import "react-toastify/dist/ReactToastify.css";
 import Pages from "../../models/Pages";
 import classes from "./login.module.scss";
-import Wrapper from "../../components/Wrapper/Wrapper";
+import Wrapper from "../../components/view/Wrapper/Wrapper";
 import LoginProps from "./interface";
 
 const Login: FC<LoginProps> = ({ user }) => {
